refactor: move configs to lachesis/
